Added testing for shells
Currently limited to command completion.
The command completion tests are handled through a shell script that is
called through the Robot Framework.  This approach was taken because
the completion tests must be handled by a shell, so coding them in
shell scripting allows for more flexibility than going through a
python layer built on top of the shell script.

Signed-off-by: Marc Khouzam <marc.khouzam@ville.montreal.qc.ca>