Use ROBOT_HELM_PATH for completion tests
If the user sets ROBOT_HELM_PATH to specify the location of the helm
binary to test, the completion tests should also use that binary.

We assume that the _dist/ dir is located one level above the location
specified by ROBOT_HELM_PATH.  This is aligned with the reality of
how Helm is build for development.

Signed-off-by: Marc Khouzam <marc.khouzam@ville.montreal.qc.ca>