Add tests for "helm repo" command
This commit adds different tests for the "helm repo" command.
It also checks that the error messages are appropriate to avoid
regressions that have been seen in the past.

Furthermore, this commit introduces the Sh class.
Having to wrap every helm command in a python method was tedious,
so the Sh class allows to use any helm command with any parameters.
In fact, the Sh class can be use to run any shell command, such as
kubectl commands or kind commands.

Signed-off-by: Marc Khouzam <marc.khouzam@ville.montreal.qc.ca>