Add tests for completion with alias or path
The dynamic completion code of Helm must be prepared to handle
when the helm to run is specified by an alias or a specific path.
These tests make sure it works as expected.

This is important as users may have both helm2 and helm3 installed
and may rely on aliases or the use of a specic path (not the PATH var)
to select which helm binary to use.  The completion must use that same
binary.

Signed-off-by: Marc Khouzam <marc.khouzam@montreal.ca>