Improve error reporting
To avoid making the user have to call a generic method at the very
end of the tests to get the final exit code, I had put the exit code
in the return call of _completionTests_verifyCompletion().

This strategy turns out to be brittle because if the test file ends
with a statement such as

if [ ! -z ${ROBOT_HELM_V3} ]; then
    _completionTests_verifyCompletion "helm comm" ""
fi

and that the execution does not enter the if, then the error code will
be the one of the if statement (success) and not the one of
_completionTests_verifyCompletion() (global test status).

So this commit falls back to the requirement of calling an exit
function at the end of the tests (_completionTests_exit()).
With this requirement, it is no longer needed for
_completionTests_verifyCompletion() to return the global test status.
Instead, it now returns the result of the current completion
verification which allows to take action on that result if so desired,
which couldn't be done before.

Signed-off-by: Marc Khouzam <marc.khouzam@montreal.ca>