Add clean target to make file
The installation of some tools has hard-code paths in them.
For example:

  $ head -1 .acceptance/.venv/bin/robot
  #!/Users/marckhouzam/git/acceptance-testing/.acceptance/.venv/bin/python3.7

If I want to move the acceptance-testing repo to a new directory on
my computer, the installed tools will no longer work due to these
hard-coded paths.  The simplest solution is to cleanup and re-install
them.

This commit as a 'make clean' target to easily remove the entire
.acceptance directory.

Signed-off-by: Marc Khouzam <marc.khouzam@montreal.ca>